[Refactor] Renamed DroneSource to Source in Drone entity
References #62.
